tests/wpt/web-platform-tests/css/css-animations/parsing/animation-delay-end-valid.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#propdef-animation-delay-end">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_valid_value("animation-delay-end", "-5ms");
test_valid_value("animation-delay-end", "0s");
test_valid_value("animation-delay-end", "10s");
test_valid_value("animation-delay-end", "20s, 10s");
</script>
